# Back/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TypeContrat": {
				"enum": [
					"CDI",
					"CDD",
					"stage"
				],
				"type": "string"
			},
			"Role": {
				"enum": [
					"Admin",
					"RH",
					"Superviseur",
					"Employe"
				],
				"type": "string"
			},
			"UserOutput": {
				"properties": {
					"id_user": {
						"type": "number",
						"format": "double"
					},
					"matricule": {
						"type": "string"
					},
					"nom": {
						"type": "string"
					},
					"prenom": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"empreinte": {
						"type": "string"
					},
					"poste": {
						"type": "string"
					},
					"type_contrat": {
						"$ref": "#/components/schemas/TypeContrat"
					},
					"date_embauche": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"date_fin_contrat": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"id_lieu": {
						"type": "number",
						"format": "double"
					},
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"id_departement": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"id_user",
					"matricule",
					"nom",
					"email",
					"phone",
					"badge",
					"poste",
					"type_contrat",
					"date_embauche",
					"id_lieu",
					"id_equipe",
					"id_departement",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"horaire": {
				"enum": [
					"jour",
					"nuit",
					"repos"
				],
				"type": "string"
			},
			"CycleDay": {
				"properties": {
					"jour": {
						"type": "number",
						"format": "double"
					},
					"shift": {
						"type": "string",
						"enum": [
							"jour",
							"nuit",
							"repos"
						]
					},
					"deb_heure": {
						"type": "string"
					},
					"fin_heure": {
						"type": "string"
					}
				},
				"required": [
					"jour",
					"shift"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPlanningEquipe": {
				"properties": {
					"id_planning": {
						"type": "number",
						"format": "double"
					},
					"debut_semaine": {
						"type": "string",
						"format": "date-time"
					},
					"jours_travail": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"horaire": {
						"$ref": "#/components/schemas/horaire"
					},
					"deb_heure": {
						"type": "string"
					},
					"fin_heure": {
						"type": "string"
					},
					"type_planning": {
						"type": "string",
						"enum": [
							"fixe",
							"cyclique"
						]
					},
					"date_debut_cycle": {
						"type": "string",
						"format": "date-time"
					},
					"cycle_pattern": {
						"items": {
							"$ref": "#/components/schemas/CycleDay"
						},
						"type": "array"
					},
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"equipe": {}
				},
				"required": [
					"id_planning",
					"type_planning",
					"id_equipe"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IPlanningEquipe.debut_semaine-or-jours_travail-or-horaire-or-deb_heure-or-fin_heure-or-id_equipe-or-type_planning-or-date_debut_cycle-or-cycle_pattern_": {
				"properties": {
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"debut_semaine": {
						"type": "string",
						"format": "date-time"
					},
					"jours_travail": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"horaire": {
						"$ref": "#/components/schemas/horaire"
					},
					"deb_heure": {
						"type": "string"
					},
					"fin_heure": {
						"type": "string"
					},
					"type_planning": {
						"type": "string",
						"enum": [
							"fixe",
							"cyclique"
						]
					},
					"date_debut_cycle": {
						"type": "string",
						"format": "date-time"
					},
					"cycle_pattern": {
						"items": {
							"$ref": "#/components/schemas/CycleDay"
						},
						"type": "array"
					}
				},
				"required": [
					"id_equipe",
					"type_planning"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"PlanningCreationParams": {
				"$ref": "#/components/schemas/Pick_IPlanningEquipe.debut_semaine-or-jours_travail-or-horaire-or-deb_heure-or-fin_heure-or-id_equipe-or-type_planning-or-date_debut_cycle-or-cycle_pattern_"
			},
			"Partial_PlanningCreationParams_": {
				"properties": {
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"debut_semaine": {
						"type": "string",
						"format": "date-time"
					},
					"jours_travail": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"horaire": {
						"$ref": "#/components/schemas/horaire"
					},
					"deb_heure": {
						"type": "string"
					},
					"fin_heure": {
						"type": "string"
					},
					"type_planning": {
						"type": "string",
						"enum": [
							"fixe",
							"cyclique"
						]
					},
					"date_debut_cycle": {
						"type": "string",
						"format": "date-time"
					},
					"cycle_pattern": {
						"items": {
							"$ref": "#/components/schemas/CycleDay"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"PlanningUpdateParams": {
				"$ref": "#/components/schemas/Partial_PlanningCreationParams_"
			},
			"EmployeePlanningForDate": {
				"properties": {
					"travaille": {
						"type": "boolean"
					},
					"shift": {
						"type": "string",
						"enum": [
							"jour",
							"nuit"
						]
					},
					"deb_heure": {
						"type": "string"
					},
					"fin_heure": {
						"type": "string"
					}
				},
				"required": [
					"travaille"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IEquipe": {
				"properties": {
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"equipe": {
						"type": "string"
					}
				},
				"required": [
					"id_equipe",
					"equipe"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_EquipeCreationParams_": {
				"properties": {
					"equipe": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"EquipeUpdateParams": {
				"$ref": "#/components/schemas/Partial_EquipeCreationParams_"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "back-rh-connect",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/users/equipe/{equipeId}": {
			"get": {
				"operationId": "GetUsersByEquipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/plannings/{id}": {
			"get": {
				"operationId": "GetPlanning",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlanningEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePlanning",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlanningEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PlanningUpdateParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePlanning",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlanningEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/plannings": {
			"post": {
				"operationId": "CreatePlanning",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlanningEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PlanningCreationParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllPlannings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPlanningEquipe"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [],
				"parameters": []
			}
		},
		"/plannings/equipe/{equipeId}": {
			"get": {
				"operationId": "GetPlanningsByEquipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPlanningEquipe"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/plannings/semaine/{debutSemaine}": {
			"get": {
				"operationId": "GetPlanningsByWeek",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPlanningEquipe"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "debutSemaine",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/plannings/equipe/{equipeId}/planning": {
			"get": {
				"operationId": "GetPlanningByEquipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlanningEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/plannings/horaire/{horaire}": {
			"get": {
				"operationId": "GetPlanningsByHoraire",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPlanningEquipe"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "horaire",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/horaire"
						}
					}
				]
			}
		},
		"/plannings/equipe/{equipeId}/date/{targetDate}": {
			"get": {
				"operationId": "GetEmployeePlanningForDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmployeePlanningForDate"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "targetDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/plannings/date/{targetDate}": {
			"get": {
				"operationId": "GetAllEmployeesPlanningForDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"equipe": {},
											"planning": {
												"$ref": "#/components/schemas/EmployeePlanningForDate"
											},
											"id_equipe": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"equipe",
											"planning",
											"id_equipe"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "targetDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/plannings/cyclique": {
			"post": {
				"operationId": "CreateCyclicPlanning",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlanningEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"date_debut": {
										"type": "string"
									},
									"equipe_type": {
										"type": "string",
										"enum": [
											"A",
											"B",
											"C"
										]
									},
									"id_equipe": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"date_debut",
									"equipe_type",
									"id_equipe"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/plannings/fixe": {
			"post": {
				"operationId": "CreateFixePlanning",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlanningEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"debut_semaine": {
										"type": "string"
									},
									"fin_heure": {
										"type": "string"
									},
									"deb_heure": {
										"type": "string"
									},
									"jours_travail": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"id_equipe": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"debut_semaine",
									"fin_heure",
									"deb_heure",
									"jours_travail",
									"id_equipe"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/plannings/actuel": {
			"get": {
				"operationId": "GetCurrentWorkingEmployees",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"equipe": {},
											"planning": {
												"$ref": "#/components/schemas/EmployeePlanningForDate"
											},
											"id_equipe": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"equipe",
											"planning",
											"id_equipe"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": []
			}
		},
		"/plannings/equipe/{equipeId}/semaine-suivante": {
			"get": {
				"operationId": "GetWeekPlanning",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"planning": {
												"$ref": "#/components/schemas/EmployeePlanningForDate"
											},
											"date": {
												"type": "string"
											}
										},
										"required": [
											"planning",
											"date"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Planning"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/equipes/{id}": {
			"get": {
				"operationId": "GetEquipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Equipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateEquipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Equipe"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EquipeUpdateParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEquipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEquipe"
								}
							}
						}
					}
				},
				"tags": [
					"Equipe"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}